Accelerate vectorizedMismatch in IL
This patch adds ArraysSupport.vectorizedMismatch as a recognized method, and
adds a SupportsInlineVectorizedMismatch flag to the code generator. This flag
is set in Z, Power, aarch64 and x86 code generator initialization if the
arraycmp opcode is supported and the TR_disableInlineVectorizedMismatch
environment variable is not set. If the flag is set, vectorizedMismatch call
nodes are transformed to a functionally equivalent tree that uses arraycmp.

Fixes: #15204
Signed-off-by: Spencer Comin <spencer.comin@ibm.com>